historico de resultados (testes e correcoes )

+++ resources/pages/atribuirresultadosprovas.php
@@ -107,13 +107,16 @@ $local =  $eventoinfo->get_local();
                               $y = 0;
                               $tamanhoatletas = count($atletas_nesta_prova);
                               do{
-                                echo "<div style= 'border: 1px solid black;  margin-left:10px; border-radius: 8px; text-align: center;'>";
-                                echo $DAO4->obter_nome_apartir_atleta_id($atletas_nesta_prova[$y]->get_idatleta());
-                                ?><br><input type="time"  <?php echo "id=time_".$provasevento[$i]->get_id()."[] name=time_".$provasevento[$i]->get_id()."[]" ?> step="1" required><br>
-                                <input type="number" <?php echo "id=lugar_".$provasevento[$i]->get_id()."[] name=lugar_".$provasevento[$i]->get_id()."[]" ?> min="1" max="100" required>
-                                <?php
-                                echo "</div>";
-                                echo "<br><br>";
+                                if(!$DAO5->historico_ja_existente($provasevento[$i]->get_id(),$atletas_nesta_prova[$y]->get_idatleta())){
+                                  //quer dizer que nao existe historico desse atleta inserido
+                                  echo "<div style= 'border: 1px solid black;  margin-left:10px; border-radius: 8px; text-align: center;'>";
+                                  echo $DAO4->obter_nome_apartir_atleta_id($atletas_nesta_prova[$y]->get_idatleta());
+                                  ?><br><input type="time"  <?php echo "id=time_".$provasevento[$i]->get_id()."[] name=time_".$provasevento[$i]->get_id()."[]" ?> step="1" required><br>
+                                  <input type="number" <?php echo "id=lugar_".$provasevento[$i]->get_id()."[] name=lugar_".$provasevento[$i]->get_id()."[]" ?> min="1" max="100" required>
+                                  <?php
+                                  echo "</div>";
+                                  echo "<br><br>";
+                                }
                                 $y++;
                               }while($y<$tamanhoatletas);
                               echo "</td>";
@@ -171,15 +174,6 @@ $local =  $eventoinfo->get_local();
             $valorb = 1;
             foreach ($resultado1 as $place){
               if($valora==$valorb){
-              /*  echo $time;
-                echo ":";
-                echo $place;
-                echo "|";
-                echo $idprova;
-                echo ";;";
-                echo $idatleta;
-                echo "//";
-                echo "<br>";*/
                 $DAO5->inserir_historico(new Historico (0,$idprova,$idatleta,$time,$place,1,true));
               }
               $valorb++;
